requirements.txt
flask
python-dotenv
python-dateutil
qrcode
twilio
edge-tts
requests
